+++ src/VisibilityState.php
@@ -1,0 +1,19 @@
+<?php
+
+// AUTOMATICALLY GENERATED.  DO NOT EDIT.
+// Use `composer build` to regenerate.
+
+namespace Wikimedia\IDLeDOM;
+
+/**
+ * VisibilityState
+ *
+ * @see https://dom.spec.whatwg.org/#enumdef-visibilitystate
+ *
+ * @phan-forbid-undeclared-magic-properties
+ */
+interface VisibilityState {
+	public const hidden = 0;
+	public const visible = 1;
+	public const prerender = 2;
+}